Updated README and moved restructure worker
