+++ src/main/java/org/radarbase/hdfs/worker/RadarHdfsRestructure.kt
@@ -14,15 +14,15 @@
  * limitations under the License.
  */
 
-package org.radarbase.hdfs
+package org.radarbase.hdfs.worker
 
 import org.apache.hadoop.conf.Configuration
 import org.apache.hadoop.fs.FileSystem
 import org.apache.hadoop.fs.Path
+import org.radarbase.hdfs.FileStoreFactory
 import org.radarbase.hdfs.accounting.Accountant
 import org.radarbase.hdfs.accounting.OffsetRange
 import org.radarbase.hdfs.path.RecordPathFactory
-import org.radarbase.hdfs.worker.RestructureWorker
 import org.slf4j.LoggerFactory
 import java.io.Closeable
 import java.io.IOException
